Refactor parser: use DocumentSyntax & add SyntaxTree/diagnostics
Rename CompilationUnitSyntax to DocumentSyntax. Introduce SyntaxTree to encapsulate DocumentSyntax, source text, and diagnostics. Add SyntaxDiagnostic infrastructure. Update parser and tests to use new structure.
