+++ src/AvroSourceGenerator.Core/Avdl/Syntax/DocumentSyntax.cs
@@ -3,12 +3,12 @@ using AvroSourceGenerator.Avdl.Syntax.Directives;
 
 namespace AvroSourceGenerator.Avdl.Syntax;
 
-public sealed record class CompilationUnitSyntax(
+public sealed record class DocumentSyntax(
     SyntaxList<IDirectiveSyntax> Directives,
     SyntaxList<IDeclarationSyntax> Declarations
 ) : ISyntaxNode
 {
-    public SyntaxKind SyntaxKind => SyntaxKind.CompilationUnit;
+    public SyntaxKind SyntaxKind => SyntaxKind.Document;
 
     public IEnumerable<ISyntaxNode> Children()
     {
