+++ src/AvroSourceGenerator.Core/Avdl/Syntax/SyntaxTree.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections.Immutable;
+using AvroSourceGenerator.Avdl.Diagnostics;
+using AvroSourceGenerator.Avdl.Text;
+
+namespace AvroSourceGenerator.Avdl.Syntax;
+
+public sealed record class SyntaxTree(SourceText SourceText, DocumentSyntax Document, ImmutableArray<SyntaxDiagnostic> Diagnostics);
